fix: Resolve circular import in validation system
This commit fixes a circular import error between `script_validator.py` and `validation_rules.py`.

The issue was caused by `script_validator.py` importing rules from `validation_rules.py`, while `validation_rules.py` imported the `ValidationResult` class from `script_validator.py`.

The fix resolves this by moving the `ValidationResult` class to its own dedicated file, `pixel_bot/core/validation_result.py`. Both modules now import this class from the new file, breaking the dependency cycle and allowing the application to start correctly.
